requirements.txt
# Core Dependencies
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# API and Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Notion Integration
notion-client==2.2.1

# Claude/AI
anthropic==0.7.0

# MCP Server
mcp==0.3.0

# CLI
click==8.1.7
typer==0.9.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
pytz==2023.3
